가게 리뷰 목록 조회

+++ src/main/java/umcspring/umc/domain/Review.java
@@ -19,7 +19,7 @@ public class Review extends BaseEntity {
     @Column(name = "review_id")
     private Long id;
 
-    private int score;
+    private Float score;
     @Column(columnDefinition = "TEXT")
     private String content;
